Moved base vendor class to an external file
